+++ solys2tracker/bodywidgets.py
@@ -258,13 +258,14 @@ class BodyTrackWidget(QtWidgets.QWidget):
         self.graph.update_plot(x_data, y_data)
         self.graph.update_headers(spec)
         with open(filename, 'w') as f:
-            print("VNIR header: {}".format(spec.fr_spectrum_header.v_header), file=f)
+            vh = spec.fr_spectrum_header.v_header
+            print("it: {}, drift: {}, VNIR header: {}".format(asdc.ITimeEnum(vh.it).to_str(), vh.drift, vh), file=f)
             s1h = spec.fr_spectrum_header.s1_header
             print("gain1: {}, offset1: {}".format(s1h.gain, s1h.offset), file=f)
             s2h = spec.fr_spectrum_header.s2_header
             print("gain2: {}, offset2: {}".format(s2h.gain, s2h.offset), file=f)
             print("", file=f)
-            # spec.to_npl_format()
+            spec.to_npl_format()
             for i in range(0, asdc.MAX_WLEN - asdc.MIN_WLEN + 1):
                 print("{:.3f}\t{:.3f}".format(i + asdc.MIN_WLEN, spec.spec_buffer[i]).replace(".",","), file=f)
             f.close()